+++ Soundboard.Tests/Mp3SoundTests.cs
@@ -32,18 +32,6 @@ namespace Soundboard.Tests
 
         [Test]
         [TestCase(@"example-sounds\mpthreetest.mp3")]
-        public void It_can_play_an_mp3_file(string pathToSoundFile)
-        {
-            var theSoundFile = GetTheProjectRelativePathForThisProjectFile(pathToSoundFile);
-
-            Subject.LoadFile(theSoundFile);
-            Subject.Play();
-
-            Subject.CurrentPlaybackState.ShouldEqual(PlaybackState.Playing);
-        }
-
-        [Test]
-        [TestCase(@"example-sounds\mpthreetest.mp3")]
         public void It_can_block_until_playback_is_complete(string pathToSoundFile)
         {
             var theSoundFile = GetTheProjectRelativePathForThisProjectFile(pathToSoundFile);
